Fix the test error due to call Connection.Server when Connection is pass in null in tests
