+++ Neo4j.Driver/Neo4j.Driver.Tests/SessionTests.cs
@@ -86,7 +86,6 @@ namespace Neo4j.Driver.Tests
                 await session.RunAsync("lalalal");
 
                 mockConn.Verify(x => x.Run("lalalal", new Dictionary<string, object>(), It.IsAny<ResultCursorBuilder>(), true), Times.Once);
-                mockConn.Verify(x => x.ReceiveOneAsync(default(CancellationToken)), Times.Once);
                 mockConn.Verify(x => x.SendAsync());
             }
